requirements.txt

streamlit


opencv-python
pillow
torch
ultralytics

numpy
pandas
matplotlib

